upadte: move away load/loadMore for messages and channels
Took 17 minutes
